Update whitelist.md
It appears that the Red Bull ad being showcased was ran (possibly inclusively) in Italy. I don't believe that it was ran recently to support Hong Kong. There are rumors about it being used to support Taiwan, but I haven't been able to find any evidence for that either. I'm inclined to think it's not intended for Taiwan or Hong Kong because the ad looks to have been run in early 2019, before the situation in Hong Kong was as developed, and years after the turn of Taiwan.

Sources:
- https://video.repubblica.it/spettacoli-e-cultura/poliziotti-contro-lo-spot-della-red-bull-ci-criminalizza-non-e-satira/329990/330591
- https://youtu.be/QIMDlCqBGVk
- https://www.youtg.net/v3/canali/in-sardegna/15255-cagliari-sindacati-di-polizia-contro-red-bull-lo-spot-istiga-all-odio-contro-le-forze-dell-ordine