Remove Marvel
The inclusion of Marvel and Disney in this list detracts from the underlying cause of this list, and is based on factually incomplete information. Scott Derrickson has gone on record calling the move to change the Ancient One to a white woman "evil, but the lesser of two evils", and doing so to avoid portraying racist stereotypes. 

https://www.express.co.uk/entertainment/films/725894/Doctor-Strange-whitewashing-Tilda-Swinton-Ancient-One-Wong-Scott-Derrickson 

Let's also not forget that Winnie the Pooh (an asset of Disney) is banned in China, yet Disney still continues to support that character. In fact, it has not actively removed Winnie the Pooh themed rides from the Shanghai Disneyland, though the Chinese government is forcefully removing all references to it

https://www.independent.co.uk/arts-entertainment/books/news/winnie-the-pooh-disneyland-china-ban-xi-jinping-meme-shanghai-president-a8660461.html